+++ example/example_app.dart
@@ -1,0 +1,8 @@
+import 'package:flutter/material.dart';
+import 'package:flutter_particle_background/flutter_particle_background.dart';
+class TestScreen extends StatelessWidget {
+  @override
+  Widget build(BuildContext context) {
+    return ParticleBackground();
+  }
+}